requirements.txt
django>=5.0.0
django-ninja>=1.1.0
groq>=0.4.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
django-cors-headers>=4.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
